Day_01_Gaussian_Noise_Basics/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.5.0
tqdm>=4.64.0
scikit-image>=0.19.0
imageio>=2.19.0
opencv-python>=4.6.0
omegaconf>=2.2.0
pytest>=7.0.0
pandas>=2.0.0